# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /**
   * @minLength 4
   * @maxLength 20
   */
  username: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UserInfo {
  id: number;
  username: string;
}

export interface AuthResponse {
  user: UserInfo;
  message: string;
}

export interface Class {
  id: number;
  userId: number;
  name: string;
  color: string;
  createdAt: string;
}

export interface CreateClassRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  color: string;
}

export type AssignmentPriority =
  (typeof AssignmentPriority)[keyof typeof AssignmentPriority];

export const AssignmentPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Assignment {
  id: number;
  userId: number;
  classId?: number | null;
  className?: string | null;
  title: string;
  notes?: string | null;
  dueDate?: string | null;
  priority: AssignmentPriority;
  completed: boolean;
  createdAt: string;
}

export type CreateAssignmentRequestPriority =
  (typeof CreateAssignmentRequestPriority)[keyof typeof CreateAssignmentRequestPriority];

export const CreateAssignmentRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateAssignmentRequest {
  classId?: number | null;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  notes?: string | null;
  dueDate?: string | null;
  priority: CreateAssignmentRequestPriority;
}

export type UpdateAssignmentRequestPriority =
  (typeof UpdateAssignmentRequestPriority)[keyof typeof UpdateAssignmentRequestPriority];

export const UpdateAssignmentRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateAssignmentRequest {
  classId?: number | null;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  notes?: string | null;
  dueDate?: string | null;
  priority?: UpdateAssignmentRequestPriority;
  completed?: boolean;
}

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType];

export const ResourceType = {
  link: "link",
  image: "image",
  note: "note",
} as const;

export interface Resource {
  id: number;
  userId: number;
  classId?: number | null;
  className?: string | null;
  type: ResourceType;
  title: string;
  content: string;
  createdAt: string;
}

export type CreateResourceRequestType =
  (typeof CreateResourceRequestType)[keyof typeof CreateResourceRequestType];

export const CreateResourceRequestType = {
  link: "link",
  image: "image",
  note: "note",
} as const;

export interface CreateResourceRequest {
  classId?: number | null;
  type: CreateResourceRequestType;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  content: string;
}