# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: classes
    description: Class management
  - name: assignments
    description: Assignment management
  - name: resources
    description: Notes and resources management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /classes:
    get:
      operationId: getClasses
      tags: [classes]
      summary: Get all classes for current user
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a new class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /classes/{id}:
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assignments:
    get:
      operationId: getAssignments
      tags: [assignments]
      summary: Get all assignments for current user
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create a new assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assignments/{id}:
    put:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentRequest"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /resources:
    get:
      operationId: getResources
      tags: [resources]
      summary: Get all resources for current user
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a new resource (image or link)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceRequest"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /resources/{id}:
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 20
        password:
          type: string
          minLength: 8
      required: [username, password]

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        message:
          type: string
      required: [user, message]

    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required: [id, username]

    Class:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, color, createdAt]

    CreateClassRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        color:
          type: string
      required: [name, color]

    Assignment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        classId:
          type: integer
          nullable: true
        className:
          type: string
          nullable: true
        title:
          type: string
        notes:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, priority, completed, createdAt]

    CreateAssignmentRequest:
      type: object
      properties:
        classId:
          type: integer
          nullable: true
        title:
          type: string
          minLength: 1
          maxLength: 200
        notes:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
      required: [title, priority]

    UpdateAssignmentRequest:
      type: object
      properties:
        classId:
          type: integer
          nullable: true
        title:
          type: string
          minLength: 1
          maxLength: 200
        notes:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        completed:
          type: boolean

    Resource:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        classId:
          type: integer
          nullable: true
        className:
          type: string
          nullable: true
        type:
          type: string
          enum: [link, image, note]
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, content, createdAt]

    CreateResourceRequest:
      type: object
      properties:
        classId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [link, image, note]
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
      required: [type, title, content]