# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentPriority } from "./assignmentPriority";

export interface Assignment {
  id: number;
  userId: number;
  classId?: number | null;
  className?: string | null;
  title: string;
  notes?: string | null;
  dueDate?: string | null;
  priority: AssignmentPriority;
  completed: boolean;
  createdAt: Date;
}