# lib/api-zod/src/generated/types/createAssignmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAssignmentRequestPriority } from "./createAssignmentRequestPriority";

export interface CreateAssignmentRequest {
  classId?: number | null;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  notes?: string | null;
  dueDate?: string | null;
  priority: CreateAssignmentRequestPriority;
}