# lib/api-zod/src/generated/types/createResourceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResourceRequestType } from "./createResourceRequestType";

export interface CreateResourceRequest {
  classId?: number | null;
  type: CreateResourceRequestType;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  content: string;
}