# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  /**
   * @minLength 4
   * @maxLength 20
   */
  username: string;
  /** @minLength 8 */
  password: string;
}