# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from "./resourceType";

export interface Resource {
  id: number;
  userId: number;
  classId?: number | null;
  className?: string | null;
  type: ResourceType;
  title: string;
  content: string;
  createdAt: Date;
}