# lib/api-zod/src/generated/types/updateAssignmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAssignmentRequestPriority } from "./updateAssignmentRequestPriority";

export interface UpdateAssignmentRequest {
  classId?: number | null;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title?: string;
  notes?: string | null;
  dueDate?: string | null;
  priority?: UpdateAssignmentRequestPriority;
  completed?: boolean;
}